chore: upgrade to node `v18.12.1` LTS (#406)
