+++ .github/workflows/build.yml
@@ -10,7 +10,7 @@ jobs:
 
         strategy:
             matrix:
-                node-version: [16.17.1]
+                node-version: [18.12.1]
 
         steps:
             - uses: actions/checkout@v3
@@ -20,7 +20,7 @@ jobs:
 
             - uses: pnpm/action-setup@v2
               with:
-                  version: 7
+                  version: 8
 
             - uses: actions/setup-node@v3
               with:
