+++ .github/workflows/deps.yml
@@ -11,7 +11,7 @@ jobs:
 
         strategy:
             matrix:
-                node-version: [16.17.1]
+                node-version: [18.12.1]
 
         steps:
             - uses: actions/checkout@v3
@@ -21,7 +21,7 @@ jobs:
 
             - uses: pnpm/action-setup@v2
               with:
-                  version: 7
+                  version: 8
 
             - uses: actions/setup-node@v3
               with:
