+++ .testcaferc.json
@@ -9,6 +9,8 @@
 	"skipJsErrors": true,
 	"selectorTimeout": 20000,
 	"assertionTimeout": 15000,
+	"appCommand": "pnpm serve",
+	"appInitDelay": 5000,
 	"compilerOptions": {
 		"typescript": {
 			"customCompilerModulePath": "../typescript",
