+++ package.json
@@ -1,6 +1,9 @@
 {
 	"name": "arkvault",
 	"version": "1.14.0",
+	"engines": {
+		"node": ">=18.12.1"
+	},
 	"private": true,
 	"description": "ARKVault",
 	"homepage": "./",
@@ -27,12 +30,12 @@
 		"format": "pnpm lint && pnpm prettier",
 		"lint": "eslint '*/**/*.{js,ts,tsx}' --quiet --fix",
 		"prettier": "prettier --write \"./*.{ts,tsx,js,json,md}\" \"./**/*.{ts,tsx,js,json,md}\"",
-		"qa": "pnpm install && pnpm build && pnpm serve -- --host=0.0.0.0",
-		"serve": "vite preview --port 5000",
+		"qa": "pnpm install && pnpm build && pnpm serve",
+		"serve": "vite preview --port 5000 --host=0.0.0.0",
 		"serve:https": "vite preview --port 5000 --https",
 		"test": "vitest --update",
 		"test:coverage": "vitest run --coverage",
-		"test:e2e": "pnpm serve & E2E_HOST='http://localhost:5000' gherkin-testcafe --config-file=.testcaferc.json",
+		"test:e2e": "E2E_HOST='http://localhost:5000' gherkin-testcafe chrome --config-file=.testcaferc.json",
 		"test:e2e:ci": "E2E_HOST='http://localhost:5000' gherkin-testcafe --config-file=.testcaferc.json chrome:headless -q attemptLimit=3,successThreshold=1",
 		"test:e2e:ci:firefox": "E2E_HOST='http://localhost:5000' gherkin-testcafe --config-file=.testcaferc.json firefox:headless -q attemptLimit=3,successThreshold=1",
 		"test:e2e:ci:edge": "E2E_HOST='http://localhost:5000' gherkin-testcafe --config-file=.testcaferc.json edge:headless -q attemptLimit=3,successThreshold=1",
@@ -61,6 +64,7 @@
 		"downshift": "6.1.12",
 		"focus-visible": "^5.2.0",
 		"framer-motion": "^9.1.7",
+		"gherkin-testcafe": "^7.0.0",
 		"i18next": "^22.4.12",
 		"locale-currency": "^0.0.2",
 		"multiformats": "^11.0.2",
@@ -72,7 +76,7 @@
 		"react-error-boundary": "^3.1.4",
 		"react-hook-form": "6.15.8",
 		"react-i18next": "^12.2.0",
-		"react-idle-timer": "4.6.4",
+		"react-idle-timer": "5.7.2",
 		"react-linkify": "^1.0.0-alpha",
 		"react-loading-skeleton": "^3.2.0",
 		"react-qr-reader": "3.0.0-beta-1",
@@ -98,7 +102,7 @@
 		"@babel/preset-env": "^7.22.2",
 		"@babel/preset-react": "^7.22.3",
 		"@babel/preset-typescript": "^7.21.0",
-		"@cucumber/cucumber": "^8.11.1",
+		"@cucumber/cucumber": "^9.3.0",
 		"@emotion/babel-plugin-jsx-pragmatic": "^0.2.0",
 		"@ledgerhq/hw-transport-mocker": "^6.27.15",
 		"@testing-library/dom": "^8.20.0",
@@ -122,7 +126,7 @@
 		"babel-plugin-macros": "^3.1.0",
 		"cross-fetch": "^3.1.5",
 		"dlv": "^1.1.3",
-		"eslint": "^8.36.0",
+		"eslint": "^8.45.0",
 		"eslint-config-prettier": "^8.7.0",
 		"eslint-plugin-import": "^2.27.5",
 		"eslint-plugin-prettier": "^4.2.1",
@@ -135,7 +139,6 @@
 		"eslint-plugin-testing-library": "^5.10.2",
 		"eslint-plugin-unicorn": "^47.0.0",
 		"eslint-plugin-unused-imports": "^2.0.0",
-		"gherkin-testcafe": "5.6.0",
 		"history": "^5.3.0",
 		"identity-obj-proxy": "^3.0.0",
 		"jest-extended": "^3.2.4",
@@ -151,17 +154,16 @@
 		"rollup": "^3.19.1",
 		"rollup-plugin-visualizer": "^5.9.0",
 		"tailwindcss": "3.3.2",
-		"testcafe": "^2.4.0",
 		"twin.macro": "2.8.2",
 		"typescript": "^5.0.4",
-		"vite": "4.3.9",
+		"vite": "4.4.7",
 		"vite-plugin-babel-macros": "^1.0.6",
 		"vite-plugin-optimize-persist": "^0.1.2",
 		"vite-plugin-package-config": "^0.1.1",
 		"vite-plugin-pwa": "^0.15.1",
 		"vite-plugin-svgr": "^3.2.0",
 		"vite-tsconfig-paths": "^4.0.7",
-		"vitest": "^0.29.3",
+		"vitest": "^0.29.8",
 		"workbox-window": "^6.6.0"
 	},
 	"babelMacros": {
