+++ src/app/hooks/use-auto-signout.ts
@@ -6,7 +6,7 @@ import { useCallback } from "react";
 export const useAutoSignOut = (profile?: Contracts.IProfile) => {
 	const history = useHistory();
 
-	const idleTimeout = 1000 * 60 * (profile?.settings().get(Contracts.ProfileSetting.AutomaticSignOutPeriod, 15) ?? 0);
+	const timeout = 1000 * 60 * (profile?.settings().get(Contracts.ProfileSetting.AutomaticSignOutPeriod, 15) ?? 1);
 
 	const onIdle = useCallback(() => {
 		if (history.location.pathname === "/") {
@@ -17,14 +17,12 @@ export const useAutoSignOut = (profile?: Contracts.IProfile) => {
 	}, [history]);
 
 	const { start, pause } = useIdleTimer({
-		crossTab: {
-			emitOnAllTabs: true,
-		},
+		crossTab: true,
 		debounce: 500,
 		onIdle,
 		startManually: true,
 		stopOnIdle: false,
-		timeout: idleTimeout,
+		timeout,
 	});
 
 	if (!profile) {
